server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "aj045045",email = "aj045045@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "beanie (>=2.0.0,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "trio (>=0.30.0,<0.31.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
